Addressed comments to simplify code and use host arg
- No longer need projectFolder or typescript args
- Sending out PR for generator-office that is tested against this branch.  Ultimately we will have to incoporate the convert-to-single-host logic to add project repos before generator-office can be updated to invoke the 'convert-to-single-host' npm script
